amount 타입 변경

See merge request s12-fintech-finance-sub1/S12P21B110!78

+++ server/src/main/java/com/b110/jjeonchongmu/domain/mypage/dto/myaccount/MyAccountResponseDto.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class MyAccountResponseDto {
     private String gatheringName;
     private String gatheringAccountNo;
-    private int gatheringAccountBalance;
+    private Long gatheringAccountBalance;
     private List<TradeDto> tradeList = new ArrayList<>();
 
     public MyAccountResponseDto(User user, List<TradeHistoryDTO> tradeHistoryDTOList) {
